requirements.txt
Flask
Flask-SQLAlchemy
requests
requests-oauthlib
python-dotenv
APScheduler
pytz
google-genai
cryptography
waitress